requirements.txt
Flask-Session
Flask==1.1.2
gunicorn==19.9.0
python-dotenv
redis
bcrypt
